+++ test/oap2/parameters/OAR025/ok-example.js
@@ -25,6 +25,16 @@ module.exports = {
       "get": {
         "responses": { "200": { "description": "no pagination ok" } }
       }
+    },
+    "/users/me": {
+      "get": {
+        "responses": { "206": { "description": "me endpoint ignored correctly" } }
+      }
+    },
+    "/users/{id}": {
+      "get": {
+        "responses": { "206": { "description": "single resource with 206 ignored correctly" } }
+      }
     }
   }
 };